Big commit here, started working on my senior project for real this time. So far I have the bare bones of my database set up, and have a rough sketch of what I want my schema to be in the DatabaseSchema.xlsx file.
I need to finish implementing the search queries, where a user can type in a common name or scientific name of a plant and it will search the list of their found plants. Also need to work on the UI for adding plants, try to get that working for now so I have something to work with. Maybe make a road map for my project as well.

So far so good though, getting a better grasp of what's actually going on. Keep it up Tylor!